lib/promscrape/discovery/yandexcloud: further code cleanup after 83a4abda3f42b784efeddd6e71766a1a8b3b9cd2
